% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{align_plots}
\alias{align_plots}
\title{Align decision tree and heatmap:}
\usage{
align_plots(
  dheat,
  dtree,
  heat_rel_height,
  show = c("heat-tree", "heat-only", "tree-only")
)
}
\arguments{
\item{dheat}{ggplot2 grob object of the heatmap.}

\item{dtree}{ggplot2 grob object of the decision tree}

\item{heat_rel_height}{Relative height of heatmap compared to whole figure (with tree).}

\item{show}{Character string indicating which components of the decision tree-heatmap
should be drawn. Can be 'heat-tree', 'heat-only' or 'tree-only'.}
}
\value{
A gtable/grob object of the decision tree (top) and heatmap (bottom).
}
\description{
Align decision tree and heatmap:
}
