% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clust.R
\name{clust_samp_func}
\alias{clust_samp_func}
\title{Performs clustering of samples.}
\usage{
clust_samp_func(leaf_node = NULL, dat, clust_vec, clust_samps)
}
\arguments{
\item{leaf_node}{Integer value indicating terminal node id.}

\item{dat}{Dataframe of the original dataset. Samples may be reordered.}

\item{clust_vec}{Character vector of variable names to be applied clustering on.
Can include class labels.}

\item{clust_samps}{if TRUE, clusters samples within each leaf_node using Gower metric
based on the values of the variables in clust_vec
and returns the ordered dataframe for each leaf_node.}
}
\value{
Dataframe of reordered original dataset when clust_samps == TRUE.
}
\description{
Performs clustering of samples.
}
