% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{diabetes}
\alias{diabetes}
\title{Diabetes patient records.}
\format{
A data frame with 768 observations and 9 variables.

http://archive.ics.uci.edu/ml/datasets/diabetes
}
\usage{
diabetes
}
\description{
Fetched from PMLB.
}
\keyword{datasets}
