% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{flags}
\alias{flags}
\title{Details of 178 nations and their flags.}
\format{
A data frame with 178 observations and 44 variables.
}
\usage{
flags
}
\description{
Fetched from PMLB.
}
\keyword{datasets}
