% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{galaxy}
\alias{galaxy}
\title{Galaxy dataset for regression.}
\format{
An object of class \code{data.frame} with 323 rows and 5 columns.
}
\usage{
galaxy
}
\description{
Fetched from PMLB.
Galaxy positions?
}
\keyword{datasets}
