% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pollen}
\alias{pollen}
\title{Synthetic pollen dataset generated by
David Coleman at RCA Laboratories in Princeton.}
\format{
A data frame with 3848 observations and 5 variables:
\code{ridge}, \code{nub}, \code{crack}, \code{weight}
and \code{target} (density).

https://www.openml.org/d/529
}
\usage{
pollen
}
\description{
Fetched from PMLB.
}
\keyword{datasets}
