% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fit.R
\name{prep_data}
\alias{prep_data}
\title{------------------------------------------------------------------------------------
Prepare dataset}
\usage{
prep_data(data, target_lab, task, feat_types = NULL)
}
\arguments{
\item{data}{Original data frame with features to be converted to correct types.}

\item{target_lab}{Name of the column in data that contains target/label information.}

\item{task}{Character string indicating the type of problem,
either 'classification' (categorical outcome) or 'regression' (continuous outcome).}

\item{feat_types}{Named vector indicating the type of each features,
e.g., c(sex = 'factor', age = 'numeric').
If feature types are not supplied, infer from column type.}
}
\value{
List of dataframes (training + test) with proper feature types and target name.
}
\description{
------------------------------------------------------------------------------------
Prepare dataset
}
