% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_feats.R
\name{prepare_feats}
\alias{prepare_feats}
\title{Prepares the feature dataframes for tiles.
If R does not recognize a categorical feature (input from user) as factor,
converts to factor.}
\usage{
prepare_feats(dat, disp_feats, feat_types, clust_feats, trans_type)
}
\arguments{
\item{dat}{Dataframe with samples from original dataset ordered according to
the clustering within each leaf node.}

\item{disp_feats}{Character vector specifying features to be displayed.}

\item{feat_types}{Named vector indicating the type of each features,
e.g., c(sex = 'factor', age = 'numeric').}

\item{clust_feats}{If TRUE, performs cluster on the features across all samples.}

\item{trans_type}{Character string specifying transformation type,
can be 'scale' or 'normalize'.}
}
\value{
a list of two dataframes (continuous and categorical)
from the original dataset.
}
\description{
Prepares the feature dataframes for tiles.
If R does not recognize a categorical feature (input from user) as factor,
converts to factor.
}
