% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.ggHeatTree}
\alias{print.ggHeatTree}
\title{Print a ggHeatTree object.
Adopted from
https://github.com/daattali/ggExtra/blob/master/R/ggMarginal.R#L207-L244.}
\usage{
\method{print}{ggHeatTree}(x, newpage = is.null(vp), vp = NULL, ...)
}
\arguments{
\item{x}{ggHeatTree (gtable grob) object.}

\item{newpage}{Should a new page (i.e., an empty page) be drawn before the
ggHeatTree is drawn?}

\item{...}{ignored}
}
\description{
\code{ggHeatTree} objects are created from \code{heat_tree()}. This is the S3
generic print method to print the result of the scatterplot with its marginal
plots.
}
\keyword{internal}
