% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{saheart}
\alias{saheart}
\title{A retrospective sample of males in a heart-disease high-risk region of the Western Cape.}
\format{
A data frame with nine variables.
}
\usage{
saheart
}
\description{
Fetched from PMLB.
}
\details{
https://rdrr.io/cran/ElemStatLearn/man/SAheart.html
https://www.rdocumentation.org/packages/BeSS/versions/1.0.6/topics/SAheart
need to check dimensions.
}
\keyword{datasets}
