% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_norm.R
\name{scale_norm}
\alias{scale_norm}
\title{Performs transformation of continuous variables for the heatmap color scales.}
\usage{
scale_norm(x, trans_type)
}
\arguments{
\item{x}{Numeric vector.}

\item{trans_type}{Character string of 'scale' or 'normalize'.
If 'scale', subtract the mean and divide by the standard deviation.
If 'normalize', i.e., max-min normalize, subtract the min and divide by the max.
More information on what transformation to choose can be acquired here:
https://cran.rstudio.com/web/packages/heatmaply/vignettes/heatmaply.html#scale}
}
\value{
Numeric vector of transformed x.
}
\description{
Performs transformation of continuous variables for the heatmap color scales.
}
