% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_tree.R
\name{term_node_pos}
\alias{term_node_pos}
\title{Determines terminal node position.}
\usage{
term_node_pos(plot_data, dat)
}
\arguments{
\item{plot_data}{Dataframe output of `ggparty:::get_plot_data()`.}

\item{dat}{Dataframe of prediction values with scaled columns
and clustered samples.}
}
\value{
Dataframe with terminal node information.
}
\description{
Create node layout using a bottom-up approach (literally) and
overwrites ggparty-precomputed positions in plot_data.
}
