% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_covid}
\alias{test_covid}
\title{External test dataset.
Medical information of Wuhan patients collected between
2020-01-10 and 2020-02-18.}
\format{

}
\usage{
test_covid
}
\description{
External test dataset.
Medical information of Wuhan patients collected between
2020-01-10 and 2020-02-18.
}
\keyword{datasets}
