% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{titanic}
\alias{titanic}
\title{Titanic.}
\format{
A data frame with 12 variables: \code{PassengerId}, \code{Survived},
  \code{Pclass}, \code{Name}, \code{Sex}, \code{Age}, \code{SipSp}, \code{Parch},
  \code{Ticket}, \code{Fare}, \code{Cabin} and \code{Embarked}.
}
\usage{
titanic
}
\description{
Titanic training dataset from titanic::titanic_train.
}
\keyword{datasets}
