% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{train_covid}
\alias{train_covid}
\title{Training dataset.
Medical information of Wuhan patients collected between
2020-01-10 and 2020-02-18.
Containing NAs.}
\format{
A data frame with 375 observations and 77 variables.

An interpretable mortality prediction model for COVID-19 patients.
Yan et al.
https://doi.org/10.1038/s42256-020-0180-7
https://github.com/HAIRLAB/Pre_Surv_COVID_19
}
\usage{
train_covid
}
\description{
Training dataset.
Medical information of Wuhan patients collected between
2020-01-10 and 2020-02-18.
Containing NAs.
}
\keyword{datasets}
