% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{waveform}
\alias{waveform}
\title{Waveform}
\format{
A data frame with 41 variables attributes describing the waveform,
all of which include noise. The latter 19 attributes are all noise attributes
with mean 0 and variance 1.

https://datahub.io/machine-learning/waveform-5000

Breiman,L., Friedman,J.H., Olshen,R.A., and Stone,C.J. (1984).
Classification and Regression Trees. Wadsworth International, pp 49-55, 169.
}
\usage{
waveform
}
\description{
Generator generating 3 classes of waves.
Each class is generated from a combination of 2 of 3 "base" waves.
}
\keyword{datasets}
