% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wine}
\alias{wine}
\title{Wine recognition dataset.}
\format{
A data frame with 14 variables.
}
\usage{
wine
}
\description{
Import with data(wine, package = 'rattle'). Dependent variable: Type.
}
\keyword{datasets}
