% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wine_quality_red}
\alias{wine_quality_red}
\title{wine_quality_red.}
\format{
A data frame with 12 variables: \code{fixed acidity}, \code{volatile acidity},
\code{citric acid}, \code{residual sugar}, \code{chlorides}, \code{free sulfur dioxide},
\code{total sulfur dioxide}, \code{density}, \code{pH}, \code{sulphates},
\code{alcohol}, \code{quality}.

http://archive.ics.uci.edu/ml/datasets/Wine+Quality
}
\usage{
wine_quality_red
}
\description{
Fetched from PMLB.
Wine quality
}
\keyword{datasets}
