% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_tree.R
\name{compute_tree}
\alias{compute_tree}
\title{Compute decision tree from data set}
\usage{
compute_tree(
  dat,
  data_test,
  task,
  clust_samps,
  clust_target,
  print_eval,
  custom_tree,
  custom_layout,
  lev_fac,
  panel_space
)
}
\arguments{
\item{dat}{Tidy dataset with dependent variable labelled 'my_target'.}

\item{data_test}{Tidy test dataset. If NULL, heatmap displays (training) `data`.}

\item{task}{Character string indicating the type of problem,
either 'classification' (categorical outcome) or 'regression' (continuous outcome).}

\item{clust_samps}{Logical. If TRUE, hierarhical clustering would be performed
among samples within each leaf node.}

\item{clust_target}{Logical. If TRUE, target/label is included in hierarchical clustering
of samples within each leaf node and might yield a more interpretable heatmap.}

\item{print_eval}{Logical. If TRUE, print evaluation of the tree performance.}

\item{custom_tree}{Custom tree with the partykit syntax.
https://cran.r-project.org/web/packages/partykit/vignettes/partykit.pdf
If NULL, a conditional inference tree is computed.}

\item{custom_layout}{Dataframe with 3 columns: id, x and y
for manually input custom layout.}

\item{lev_fac}{Relative weight of child node positions
according to their levels, commonly ranges from 1 to 1.5.
1 for parent node perfectly in the middle of child nodes.}

\item{panel_space}{Spacing between facets relative to viewport,
recommended to range from 0.001 to 0.01.}
}
\value{
A list of results from `partykit::ctree`, smart layout data,
terminal data, node labels, and features to be displayed.
}
\description{
Compute decision tree from data set
}
