% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{diabetes}
\alias{diabetes}
\title{Diabetes patient records.}
\format{
A data frame with 768 observations and 9 variables:
\code{Pregnancies}, \code{Glucose}, \code{BloodPressure}, \code{SkinThickness}, \code{Insulin},
\code{BMI}, \code{DiabetesPedigreeFunction}, \code{Age} and \code{Outcome}.
}
\usage{
diabetes
}
\description{
http://archive.ics.uci.edu/ml/datasets/diabetes
https://www.kaggle.com/uciml/pima-indians-diabetes-database
}
\keyword{datasets}
