% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_heat.R
\name{draw_heat}
\alias{draw_heat}
\title{Draws the heatmap to be placed below the decision tree.}
\usage{
draw_heat(
  dat,
  feat_names,
  disp_feats,
  class_cols,
  panel_space,
  feat_types = NULL,
  trans_type = "normalize",
  cont_cols = ggplot2::scale_fill_viridis_c(),
  cate_cols = ggplot2::scale_fill_viridis_d(option = "D", begin = 0.3, end = 0.9),
  clust_feats = TRUE,
  class_space = 0.03,
  class_pos = "top"
)
}
\arguments{
\item{dat}{Dataframe with samples from original dataset ordered according to
the clustering within each leaf node.}

\item{feat_names}{Character vector specifying the feature names in dat.}

\item{disp_feats}{Character vector specifying features to be displayed.}

\item{class_cols}{Vector of RGBs for the class colors,
defaults to a colorblind friendly palette.}

\item{panel_space}{Spacing between facets relative to viewport,
recommended to range from 0.001 to 0.01.}

\item{feat_types}{Named vector indicating the type of each features,
e.g., c(sex = 'factor', age = 'numeric').
If feature types are not supplied, infer from column type.}

\item{trans_type}{Character string specifying transformation type,
can be 'scale' or 'normalize'.}

\item{cont_cols}{Function determine color scale for continuous variable,
defaults to viridis option D.}

\item{cate_cols}{Function determine color scale for nominal categorical variable,
defaults to viridis option D.}

\item{clust_feats}{If TRUE, performs cluster on the features.}

\item{class_space}{Numeric value indicating spacing between
the class label and the rest of the features}

\item{class_pos}{Character string specifying the position of the class label
on heatmap, can be 'top', 'bottom' or 'none'.}
}
\value{
A ggplot2 grob object of the heatmap.
}
\description{
Draws the heatmap to be placed below the decision tree.
}
