% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_heat.R
\name{draw_heat}
\alias{draw_heat}
\title{Draws the heatmap.}
\usage{
draw_heat(
  dat,
  feat_names,
  disp_feats,
  target_cols,
  panel_space,
  feat_types = NULL,
  trans_type = "normalize",
  cont_cols = ggplot2::scale_fill_viridis_c(),
  cate_cols = ggplot2::scale_fill_viridis_d(option = "D", begin = 0.3, end = 0.9),
  clust_feats = TRUE,
  target_space = 0.03,
  target_pos = "top",
  target_lab_disp = NULL
)
}
\arguments{
\item{dat}{Dataframe with samples from original dataset ordered according to
the clustering within each leaf node.}

\item{feat_names}{Character vector specifying the feature names in dat.}

\item{disp_feats}{Character vector specifying features to be displayed.}

\item{target_cols}{Function determine color scale for target,
defaults to viridis option B.}

\item{panel_space}{Spacing between facets relative to viewport,
recommended to range from 0.001 to 0.01.}

\item{feat_types}{Named vector indicating the type of each features,
e.g., c(sex = 'factor', age = 'numeric').
If feature types are not supplied, infer from column type.}

\item{trans_type}{Character string specifying transformation type,
can be 'scale' or 'normalize'.}

\item{cont_cols}{Function determine color scale for continuous variable,
defaults to viridis option D.}

\item{cate_cols}{Function determine color scale for nominal categorical variable,
defaults to viridis option D.}

\item{clust_feats}{Logical. If TRUE, performs cluster on the features.}

\item{target_space}{Numeric value indicating spacing between
the class label and the rest of the features}

\item{target_pos}{Character string specifying the position of the class label
on heatmap, can be 'top', 'bottom' or 'none'.}

\item{target_lab_disp}{Character string for displaying the label of class label
if it differs from target_lab.}
}
\value{
A ggplot2 grob object of the heatmap.
}
\description{
Draws the heatmap to be placed below the decision tree.
}
