% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_tree.R
\name{draw_tree}
\alias{draw_tree}
\title{Draws the conditional decision tree.}
\usage{
draw_tree(
  fit,
  layout,
  term_dat,
  target_cols,
  tree_space_top,
  tree_space_bottom,
  par_node_vars,
  terminal_vars,
  edge_vars,
  edge_text_vars,
  print_eval,
  x_eval,
  y_eval,
  text_eval
)
}
\arguments{
\item{fit}{ggparty object as output from partykit::ctree()}

\item{layout}{Dataframe of layout of all nodes, must include these columns:
id, x, y and y_hat.}

\item{term_dat}{Dataframe for terminal nodes, must include these columns:
id, x, y and y_hat.}

\item{target_cols}{Function determine color scale for target,
defaults to viridis option B.}

\item{tree_space_top}{Numeric value to pass to expand for top margin of tree.}

\item{tree_space_bottom}{Numeric value to pass to expand for bottom margin of tree.}

\item{par_node_vars}{Named list containing arguments to be passed to the
`geom_node_label()` call for non-terminal nodes.}

\item{terminal_vars}{Named list containing arguments to be passed to the
`geom_node_label()` call for terminal nodes.}

\item{edge_vars}{Named list containing arguments to be passed to the
`geom_edge()` call for tree edges.}

\item{edge_text_vars}{Named list containing arguments to be passed to the
`geom_edge_label()` call for tree edge annotations.}

\item{print_eval}{Logical. If TRUE, print evaluation of the tree performance.}

\item{x_eval}{Numeric value indicating x position to print performance statistics.}

\item{y_eval}{Numeric value indicating y position to print performance statistics.}

\item{text_eval}{Character string of the decision tree evaluation.}
}
\value{
A ggplot2 grob object of the decision tree.
}
\description{
Draws the conditional decision tree output from partykit::ctree(),
utilizing ggparty geoms: geom_edge, geom_edge_label, geom_node_label.
}
