% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_tree.R
\name{draw_tree}
\alias{draw_tree}
\title{Draws the conditional decision tree.}
\usage{
draw_tree(
  fit,
  class_cols,
  layout,
  term_dat,
  tree_space_top = 0.05,
  tree_space_bottom = 0.035,
  par_node_vars = list(label.size = 0, label.padding = ggplot2::unit(0.15, "lines"),
    line_list = list(ggplot2::aes(label = splitvar)), line_gpar = list(list(size = 9))),
  terminal_vars = list(label.padding = ggplot2::unit(0.25, "lines"), size = 3),
  edge_vars = list(color = "grey70", size = 0.5),
  edge_text_vars = list(color = "grey30", size = 3)
)
}
\arguments{
\item{fit}{ggparty object as output from partykit::ctree()}

\item{class_cols}{Vector of RGBs for the class colors,
defaults to a colorblind friendly palette.}

\item{layout}{Dataframe of layout of all nodes, must include these columns:
id, x, y and y_hat.}

\item{term_dat}{Dataframe for terminal nodes, must include these columns:
id, x, y and y_hat.}

\item{tree_space_top}{Numeric value to pass to expand for top margin of tree.}

\item{tree_space_bottom}{Numeric value to pass to expand for bottom margin of tree.}

\item{par_node_vars}{Named list containing arguments to be passed to the
`geom_node_label()` call for non-terminal nodes.}

\item{terminal_vars}{Named list containing arguments to be passed to the
`geom_node_label()` call for terminal nodes.}

\item{edge_vars}{Named list containing arguments to be passed to the
`geom_edge()` call for tree edges.}

\item{edge_text_vars}{Named list containing arguments to be passed to the
`geom_edge_label()` call for tree edge annotations.}
}
\value{
A ggplot2 grob object of the decision tree.
}
\description{
Draws the conditional decision tree output from partykit::ctree(),
utilizing ggparty geoms: geom_edge, geom_edge_label, geom_node_label.
}
