% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{flags}
\alias{flags}
\title{Flags.}
\format{
An object of class \code{data.frame} with 178 rows and 44 columns.
}
\usage{
flags
}
\description{
Fetched from PMLB.
Contains details of 178 nations and their flags with 43 variables.
Unsure what 'target' indicates atm.
}
\keyword{datasets}
