% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{galaxy}
\alias{galaxy}
\title{Galaxy dataset for regression.}
\format{
An object of class \code{data.frame} with 323 rows and 5 columns.
}
\usage{
galaxy
}
\description{
Fetched from PMLB.
}
\details{
#' @format A data frame with 323 observations and 5 variables:
\code{eastwest}, \code{northsouth}, \code{angle}, \code{radialposition}
and \code{target} (velocity).

https://www.openml.org/d/690
}
\keyword{datasets}
