% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_cols}
\alias{get_cols}
\title{Get color functions from character vectors}
\usage{
get_cols(my_cols, task)
}
\arguments{
\item{my_cols}{Character vectors of different hex values}

\item{task}{Character string indicating the type of problem,
either 'classification' (categorical outcome) or 'regression' (continuous outcome).}
}
\description{
Get color functions from character vectors
}
