% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mushroom}
\alias{mushroom}
\title{Mushroom records.}
\format{
A data frame with 23 variables.
}
\usage{
mushroom
}
\description{
Fetched from PMLB.
}
\details{
http://archive.ics.uci.edu/ml/datasets/Mushroom
The Audubon Society Field Guide to North American Mushrooms (1981).
G. H. Lincoff (Pres.), New York: Alfred A. Knopf
}
\keyword{datasets}
