% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pollen}
\alias{pollen}
\title{Pollen dataset for regression.}
\format{
An object of class \code{data.frame} with 3848 rows and 5 columns.
}
\usage{
pollen
}
\description{
Fetched from PMLB.
}
\keyword{datasets}
