% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_tree.R
\name{prep_data}
\alias{prep_data}
\title{Prepare dataset}
\usage{
prep_data(
  data,
  data_test = NULL,
  target_lab,
  task,
  feat_types = NULL,
  label_map = NULL
)
}
\arguments{
\item{data}{Tidy dataset.}

\item{data_test}{Tidy test dataset. If NULL, heatmap displays (training) `data`.}

\item{target_lab}{Name of the column in data that contains target/label information.}

\item{task}{Character string indicating the type of problem,
either 'classification' (categorical outcome) or 'regression' (continuous outcome).}

\item{feat_types}{Named vector indicating the type of each features,
e.g., c(sex = 'factor', age = 'numeric').
If feature types are not supplied, infer from column type.}

\item{label_map}{Named vector of the meaning of the target values,
e.g., c(`0` = 'Edible', `1` = 'Poisonous').}
}
\value{
List of dataframes (traing + test) with proper feature types and target name.
}
\description{
Prepare dataset
}
