% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{saheart}
\alias{saheart}
\title{A retrospective sample of males in a heart-disease high-risk region of the Western Cape.}
\format{
A data frame with 178 observations and 10 variables:
\code{Sbp}, \code{Tobacco}, \code{Ldl}, \code{Adiposity}, \code{Famhist}, \code{Typea},
\code{Obesity}, \code{Alcohol}, \code{Age} and \code{target}.
}
\usage{
saheart
}
\description{
Fetched from PMLB.
}
\details{
https://rdrr.io/cran/ElemStatLearn/man/SAheart.html
https://www.rdocumentation.org/packages/BeSS/versions/1.0.6/topics/SAheart
}
\keyword{datasets}
