% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_covid}
\alias{test_covid}
\title{External test dataset.
Medical information of Wuhan patients collected between
2020-01-10 and 2020-02-18.}
\format{
A data frame with 110 observations and 7 XGBoost-selected variables:
\code{PATIENT_ID}, \code{Lactate dehydrogenase},
\code{High sensitivity C-reactive protein}, \code{(\%)lymphocyte},
\code{Admission time}, \code{Discharge time} and \code{outcome}.

An interpretable mortality prediction model for COVID-19 patients.
Yan et al.
https://doi.org/10.1038/s42256-020-0180-7
https://github.com/HAIRLAB/Pre_Surv_COVID_19
}
\usage{
test_covid
}
\description{
External test dataset.
Medical information of Wuhan patients collected between
2020-01-10 and 2020-02-18.
}
\keyword{datasets}
