% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{titanic}
\alias{titanic}
\title{Fate of passengers on the fatal maiden voyage of the ocean liner *Titanic*.}
\format{
A data frame with 891 observations and 12 variables: \code{PassengerId},
\code{Pclass}, \code{Name}, \code{Sex}, \code{Age}, \code{SipSp}, \code{Parch},
\code{Ticket}, \code{Fare}, \code{Cabin}, \code{Embarked}
and \code{Survived} (target).

https://www.rdocumentation.org/packages/titanic/versions/0.1.0
}
\usage{
titanic
}
\description{
Titanic training dataset from titanic::titanic_train.
}
\keyword{datasets}
