% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wine}
\alias{wine}
\title{Results of a chemical analysis of wines grown in a specific area of Italy.}
\format{
A data frame with 178 observations and 14 variables:
\code{Alcohol}, \code{Malic}, \code{Ash}, \code{Alcalinity},
\code{Magnesium}, \code{Phenols}, \code{Flavanoids}, \code{Nonflavanoids},
\code{Proanthocyanins}, \code{Color}, \code{Hue}, \code{Dilution}, \code{Proline}
and \code{Type} (target).
}
\usage{
wine
}
\description{
Three types of wine are represented in the 178 samples,
with the results of 13 chemical analyses recorded for each sample.
}
\details{
Import with data(wine, package = 'rattle'). Dependent variable: Type.
https://rdrr.io/cran/rattle.data/man/wine.html
http://archive.ics.uci.edu/ml/datasets/wine
}
\keyword{datasets}
