% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wine_quality_red}
\alias{wine_quality_red}
\title{Red variant of the Portuguese "Vinho Verde" wine.}
\format{
A data frame with 1599 observations and 12 variables:
\code{fixed.acidity}, \code{volatile.acidity},
\code{citric.acid}, \code{residual.sugar}, \code{chlorides}, \code{free.sulfur.dioxide},
\code{total.sulfur.dioxide}, \code{density}, \code{pH}, \code{sulphates},
\code{alcohol} and \code{target} (quality).

http://archive.ics.uci.edu/ml/datasets/Wine+Quality

P. Cortez, A. Cerdeira, F. Almeida, T. Matos and J. Reis.
Modeling wine preferences by data mining from physicochemical properties.
In Decision Support Systems, Elsevier, 47(4):547-553, 2009.
}
\usage{
wine_quality_red
}
\description{
Fetched from PMLB.
Physicochemical and quality of wine.
}
\keyword{datasets}
